2023年01月11日更新：驾驶员、车辆、企业保险资料导入、统计列表等功能提交接口更新；

+++ blade-service-api/blade-anbiao-api/src/main/java/org/springblade/anbiao/cheliangguanli/entity/DeptBaoxianMingxi.java
@@ -118,7 +118,7 @@ public class DeptBaoxianMingxi implements Serializable {
      * 保司金额
      */
     @ApiModelProperty(value = "保司金额")
-    private Long avbmCompanyAmount;
+    private BigDecimal avbmCompanyAmount;
     /**
      * 序号
      */
