+++ blade-service-api/blade-anbiao-api/src/main/java/org/springblade/anbiao/cheliangguanli/entity/VehicleBaoxianMingxi.java
@@ -120,7 +120,7 @@ public class VehicleBaoxianMingxi implements Serializable {
      * 保司金额
      */
     @ApiModelProperty(value = "保司金额")
-    private Long avbmCompanyAmount;
+    private BigDecimal avbmCompanyAmount;
     /**
      * 序号
      */
